7/12 1:00

+++ src/main/java/project/fashion/entity/Invoice.java
@@ -15,7 +15,7 @@ public class Invoice {
     private String name;
     private String phone;
     private String address;
-    private Double totalAmount;
+    private Integer totalAmount;
     private String note;
     @Temporal(TemporalType.TIMESTAMP)
     private LocalDateTime createdAt;
